agent-orchestrator/requirements.txt
pulsar-client==3.3.0
kubernetes==28.1.0